requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
idna==2.10
paramiko==2.7.2
pycparser==2.20
PyNaCl==1.4.0
PyYAML==5.4.1
requests==2.25.1
scp==0.13.3
six==1.15.0
urllib3==1.26.4
setuptools~=54.2.0
click~=7.1.2
tabulate~=0.8.9